NVM IT WORKS, MY WIRING IRL WAS JUST WRONG

+++ src/main.rs
@@ -112,8 +112,9 @@ async fn servo_task(mut servo_rx: mpsc::UnboundedReceiver<ServoPosition>, servo_
                 ServoPosition::Center => (MAX_PULSE_WIDTH_US + MIN_PULSE_WIDTH_US) / 2,
                 ServoPosition::Right => MAX_PULSE_WIDTH_US,
                 ServoPosition::Custom(pos) => {
-                    (pos * (MAX_PULSE_WIDTH_US - MIN_PULSE_WIDTH_US) as f32 / 2.) as u64
-                        + (MAX_PULSE_WIDTH_US + MIN_PULSE_WIDTH_US) / 2
+                    ((pos * (MAX_PULSE_WIDTH_US - MIN_PULSE_WIDTH_US) as f32 / 2.)
+                        + (MAX_PULSE_WIDTH_US + MIN_PULSE_WIDTH_US) as f32 / 2.)
+                        as u64
                 }
             };
             servo_pwm
